Tests/LibWeb/Text/expected/hit_testing/content-editable-empty.txt

<DIV id="div1">
<DIV id="div2">
